Clean up frontend and fix bug with next episode
Getting the next episode or current episode to play should not
return an episode that has been completely played

+++ src/main/java/com/podcase/model/Episode.java
@@ -100,7 +100,7 @@ import com.podcase.serializer.PodcastSerializer;
 	            @ColumnResult(name="guid", type=String.class),
 	            @ColumnResult(name="publication_date", type=Date.class),
 	            @ColumnResult(name="retrieved_date", type=Date.class),
-	            @ColumnResult(name="play_length", type=BigInteger.class),
+	            @ColumnResult(name="play_length", type=Integer.class),
 	            @ColumnResult(name="duration", type=Integer.class),
 	            @ColumnResult(name="play_state_id", type=BigInteger.class)
 	        }
